# src/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Personal Finance Tracker API",
        "version": "1.0.0",
        "description": "API for managing personal finances including user authentication via OAuth, transaction management,\ncategory management, budgeting, and reporting. Role-based access is enforced so that users can only access their own data.\n",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "description": "Local host testing",
            "url": "http://localhost:8080"
        },
        {
            "description": "Render live server",
            "url": "https://personal-finance-tracker-api-4n2c.onrender.com"
        }
    ],
...
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieve a list of user transactions",
                "operationId": "getTransactions",
                "description": "Retrieve all transactions for the authenticated user.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            },
            "post": {
                "tags": [
                    "transaction"
                ],
                "summary": "Create a new transaction",
                "operationId": "createTransaction",
                "description": "Create a new financial transaction.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieve a single transaction by ID",
                "operationId": "getTransactionById",
                "description": "Retrieve a specific transaction for the authenticated user.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            },
            "put": {
                "tags": [
                    "transaction"
                ],
                "summary": "Update an existing transaction",
                "operationId": "updateTransaction",
                "description": "Modify the details of a transaction for the authenticated user.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input."
                    },
                    "404": {
                        "description": "Transaction not found."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            },
            "delete": {
                "tags": [
                    "transaction"
                ],
                "summary": "Delete a transaction",
                "operationId": "deleteTransaction",
                "description": "Remove a transaction for the authenticated user.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted successfully."
                    },
                    "404": {
                        "description": "Transaction not found."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
...
        "/budgets": {
            "get": {
                "tags": [
                    "budget"
                ],
                "summary": "Retrieve current budget setups for a user",
                "operationId": "getBudgets",
                "description": "Retrieve all budget records for the authenticated user.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Budgets retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Budget"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            },
            "post": {
                "tags": [
                    "budget"
                ],
                "summary": "Create a new budget",
                "operationId": "createBudget",
                "description": "Allows authenticated users to create a new budget record.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BudgetInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Budget created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BudgetInput"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
        "/budgets/{id}": {
            "put": {
                "tags": [
                    "budget"
                ],
                "summary": "Update an existing budget",
                "operationId": "updateBudget",
                "description": "Allows authenticated users to update their budget by ID.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BudgetInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Budget updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BudgetInput"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input."
                    },
                    "404": {
                        "description": "Budget not found."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            },
            "delete": {
                "tags": [
                    "budget"
                ],
                "summary": "Delete a budget",
                "operationId": "deleteBudget",
                "description": "Allows authenticated users to delete a budget by ID.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Budget deleted successfully."
                    },
                    "404": {
                        "description": "Budget not found."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "Transaction": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "trans123"
                    },
                    "userId": {
                        "type": "string",
                        "example": "user123"
                    },
                    "amount": {
                        "type": "number",
                        "example": 100.5
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-07T12:00:00Z"
                    },
                    "category": {
                        "type": "string",
                        "example": "Groceries"
                    },
                    "description": {
                        "type": "string",
                        "example": "Weekly grocery shopping"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "example": "Credit Card"
                    },
                    "transactionType": {
                        "type": "string",
                        "example": "Expense"
                    }
                },
                "required": [
                    "id",
                    "userId",
                    "amount",
                    "date",
                    "category",
                    "transactionType"
                ]
            },
            "TransactionInput": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number",
                        "example": 100.5
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-07T12:00:00Z"
                    },
                    "category": {
                        "type": "string",
                        "example": "Groceries"
                    },
                    "description": {
                        "type": "string",
                        "example": "Weekly grocery shopping"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "example": "Credit Card"
                    },
                    "transactionType": {
                        "type": "string",
                        "example": "Expense"
                    }
                },
                "required": [
                    "amount",
                    "date",
                    "category",
                    "transactionType"
                ]
            },
...
            "Budget": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "budget123"
                    },
                    "userId": {
                        "type": "string",
                        "example": "user123"
                    },
                    "category": {
                        "type": "string",
                        "example": "Groceries"
                    },
                    "amount": {
                        "type": "number",
                        "example": 500
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-01"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-31"
                    }
                },
                "required": [
                    "id",
                    "userId",
                    "category",
                    "amount",
                    "startDate",
                    "endDate"
                ]
            },
            "BudgetInput": {
                "type": "object",
                "properties": {
                    "category": {
                        "type": "string",
                        "example": "Food"
                    },
                    "amount": {
                        "type": "number",
                        "example": 500
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-01"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-31"
                    }
                },
                "required": [
                    "category",
                    "amount",
                    "startDate",
                    "endDate"
                ]
            },
...
        }
    }
}